Ensure all descriptor names are consistent
